+++ addOns/pscanrulesBeta/CHANGELOG.md
@@ -3,10 +3,11 @@ All notable changes to this add-on will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## Unreleased
+## [23] - 2020-11-18
 ### Changed
-- Update RE2/J library to latest version (1.4).
+- Update RE2/J library to latest version (1.5).
 - Maintenance changes.
+- Content Security Policy header missing scan rule changed to Medium risk in order to align with other CSP findings, and confidence to High (Issue 6301).
 
 ## [22] - 2020-06-01
 ### Added
@@ -166,6 +167,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Updated to support new addon format
 
+[23]: https://github.com/zaproxy/zap-extensions/releases/pscanrulesBeta-v23
 [22]: https://github.com/zaproxy/zap-extensions/releases/pscanrulesBeta-v22
 [21]: https://github.com/zaproxy/zap-extensions/releases/pscanrulesBeta-v21
 [20]: https://github.com/zaproxy/zap-extensions/releases/pscanrulesBeta-v20